==> Ranalysis/scripts/1_PI_curves.Rmd <==
---
title: "Photosynthesis and respiration rate calculations"
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## install packages if you dont already have them in your library
if (!require("devtools")) install.packages("devtools")
if (!require("furrr")) install.packages("furrr")
if (!require("future")) install.packages("future")
if (!require("tidyverse")) install.packages("tidyverse")
if (!require("lubridate")) install.packages("lubridate")
if (!require("cowplot")) install.packages("cowplot")
if (!require("LoLinR")) install_github('colin-olito/LoLinR') 

## load libraries
library(devtools)
library(LoLinR)
library(tidyverse)
library(lubridate)
library(cowplot)
library(broom)

## libraries for parallel processing
library(future)
library(furrr)
```

## Import metadata
```{r, warning = FALSE}
path.p <- "data/PR/pi_curves/" #the location of all your respirometry files 

# List data files
file.names <- list.files(path = path.p, pattern = "csv$")  # list all csv file names in the folder
file.names <- file.names[!grepl("metadata", file.names)]   # omit metadata from files to be read in as data

# Load PI curve sample metadata (i.e., which corals were in which runs)
sample.info <- read_csv(file = "data/PR/pi_curves/pi_curves_sample_metadata.csv")

# Load PI curve run metadata (i.e., light levels and interval times for each run)
run.info <- read_csv(file = "data/PR/pi_curves/pi_curves_run_metadata.csv")

# Join all coral and run metadata
metadata <- full_join(sample.info, run.info) %>%
  mutate(Date = as_date(as.character(Date), format = "%Y%m%d", tz = "Tahiti"))

# Select only certain columns
metadata <- metadata %>%
  select(colony_id, Run, Chamber.Vol.L, Date, Start.time, Stop.time, Light_Value, Light_Level)
```


# Read in all data files
```{r, results="hide", message=FALSE}
df <- tibble(file.name = file.names) %>%
  mutate(colony_id = gsub("_.*", "", file.name),                              # Get colony_id from filename
          info = map(colony_id, ~filter(metadata, colony_id == .)),           # Get associated sample info
         data0 = map(file.name, ~read_csv(file.path(path.p, .), skip=0, col_types = cols(.default = "d", Time = "t"))))   # Get associated O2 data

# Select only Time, Value, and Temp columns from O2 data
df <- df %>%
  mutate(data0 = map(data0, ~select(., Time, Value, Temp)))%>%
  mutate(data0 = map(data0, ~(.x %>% filter(complete.cases(.))))) #remove NAs to get rid of artifact line in our data

```

## Use the time breaks in the sample info to link O2 data with light levels
```{r, warning = FALSE}
df <- df %>%
  mutate(intervals = map2(data0, info, function(.x, .y) {
    split(.x, f = cut(as.numeric(.x$Time), breaks = as.numeric(c(.y$Start.time, last(.y$Stop.time))),
                      labels = as.character(.y$Light_Value)))})) %>%
  mutate(data = map(intervals, ~ unnest(tibble(.), .id = "Light_Value")))

## 'data' now contains the O2 data with the corresponding light value as another column
## Example of what 'data' for each sample looks like:
# df$data[[4]]
```

### Thin data
```{r, fig.height = 8, fig.width = 8}
# Set thinning parameter
thin_par <- 20

# Thin data for all samples
df <- df %>%
  mutate(thin_data = map(data, ~ slice(., seq(1, nrow(.), thin_par))))

# Create plots for full dataset and thinned data
df <- df %>%
  mutate(data_plot = map2(data, colony_id, ~ ggplot(.x, aes(x = Time, y = Value)) + 
                            facet_wrap(~ as.numeric(Light_Value), scales = "free") +
                            geom_point() +
                            labs(title = .y)),
    thin_data_plot = map2(thin_data, colony_id, ~ ggplot(.x, aes(x = Time, y = Value)) + 
                            facet_wrap(~ as.numeric(Light_Value), scales = "free") +
                            geom_point() +
                            labs(title = .y)))

# Example of plots
cowplot::plot_grid(df$data_plot[[2]], df$thin_data_plot[[2]], nrow = 2,
                   labels = c("Example plot: all data", "Example plot: thinned data"))
```

#### The full or thinned data plot for any sample can be accessed like this:
```
df %>%
  filter(colony_id == "ACR-185") %>%
  pull(thin_data_plot)
```

# Fit regressions to each interval for each sample
```{r}
# Define function for fitting LoLinR regressions to be applied to all intervals for all samples
fit_reg <- function(df) {
  rankLocReg(xall = as.numeric(df$Time), yall = df$Value, 
             alpha = 0.2, method = "pc", verbose = FALSE)
}

# Setup for parallel processing
future::plan(multisession)

# Map LoLinR function onto all intervals of each sample's thinned dataset
df <- df %>%
  mutate(regs = furrr::future_map(thin_data, function(.) {       # future_map executes function in parallel
    group_by(., Light_Value) %>%
    do(rankLcRg = fit_reg(.))
  }))

## Now 'regs' contains the fitted local regressions for each interval of each sample's thinned dataset

# Define function to pull out and plot regression diagnostics
plot_rankLcRg <- function(colony_id, interval_number) {
  df %>%
    filter(colony_id == colony_id) %>%
    pluck("regs", 1, "rankLcRg", interval_number) %>%
    plot()
}
```

### Extract slope of best regression for each interval for each sample
```{r}
df.out <- df %>% 
  unnest(regs) %>%
  mutate(micromol.L.s = map_dbl(rankLcRg, ~ pluck(., "allRegs", "b1", 1)))
```

# Adjust by chamber volume and normalize to surface area
```{r}
### Merge rates with sample info
pr <- left_join(
  select(df.out,colony_id, Light_Value, micromol.L.s),
  distinct(metadata,colony_id, Run, Chamber.Vol.L))

# Correct for chamber volume and blanks

### generate a key for the blank id
pr <- pr %>%
  mutate(Run_blank_light=paste0(Run,"_","blank","_",Light_Value))%>%
  mutate(blank_id=paste0(Run,"_","blank"))

# Get blank values -- average for each run and light value in case multiple blanks
blanks <- pr %>%
  filter(grepl("BK", colony_id)) %>%
  mutate(micromol.s.blank = micromol.L.s * Chamber.Vol.L)%>%
  group_by(Run, Light_Value) %>%
  mutate(blank_id=paste0(Run,"_","blank"))%>%
  mutate(Run_blank_light=paste0(Run,"_","blank","_",Light_Value))

blanks %>% ggplot(aes(x=as.numeric(Light_Value), y=micromol.s.blank,colour = as.factor(Run)))+
  geom_point()

pr <- pr %>% mutate(micromol.s = micromol.L.s * Chamber.Vol.L)

pr  <- dplyr::inner_join(
   pr ,
   dplyr::select(blanks, -any_of(names(pr)), Run_blank_light),
   by = "Run_blank_light"
)

#Join blank values with rest of data and subtract values from samples for same run and light value
pr <- pr %>%
  mutate(micromol.s.adj = micromol.s - micromol.s.blank) %>%
  # After correcting for blank values, remove blanks from data
  filter(!grepl("BK", colony_id))

# Import surface area data
sa <- read.csv("output/1_surface_area.csv")

# Join surface area with rest of data
pr <- left_join(pr, select(sa, colony_id, surface.area.cm2))

# Normalize rates by surface area
pr <- pr %>%
  mutate(micromol.cm2.s = micromol.s.adj / surface.area.cm2,
         micromol.cm2.h = micromol.cm2.s * 3600)
```

# Plot rates vs. irradiance for each sample
```{r, fig.height=15, fig.width = 10}
ggplot(pr, aes(x = as.numeric(Light_Value.x), y = micromol.cm2.h)) +
  geom_point(size = 3) +
  facet_wrap(~colony_id, ncol = 3)
```

# Define data  

```{r}
#specify data
Data <- pr
Data$PAR <- as.numeric(Data$Light_Value.x)
Data$Pc <- as.numeric(Data$micromol.cm2.h)
```


# Define PI curve function as a nonlinear Least Squares regression of a quadratic fit, test nls fit
#Pc ~ (Am*((AQY*PAR)/(sqrt(Am^2 + (AQY*PAR)^2)))-Rd), data=., start=list(Am=(max(.$Pc)-min(.$Pc)),  AQY=0.001, Rd=-min(.$Pc))
Aquatic Photosynthesis, Falkowski 
PAR = irradiance from 400-700nm (also called I or E)
PC = oxygen flux rate
Pmax = max photosynthesis (also called Am)  
alpha = quantum yeild (also called AQY)  
I or E or PAR = irradiance 
Rd = dark respiration
Ik (saturating irradiance) is the point at which photosynthesis reaches the max of initial slope = Am/AQY
Ic=(Am*Rd)/(AQY*(sqrt(Am^2-Rd^2)))
Equation for Ic derived from quadratic equation above. Ic = Par when Pc = 0 (x intercept). Ic = light compensation point; point at which photosynthesis is released from carbon limitation. 

Run nls model 
Using flexible initial values based on input data:  
```{r}
nls_data <- Data %>% 
   group_by(colony_id) %>%
   nest(-colony_id) %>%
   mutate(model1 = map(data, ~ 
                         nls(Pc ~ (Am*((AQY*PAR)/(sqrt(Am^2 + (AQY*PAR)^2)))-Rd), data=., start=list(Am=(max(.$Pc)-min(.$Pc)),  AQY=0.001, Rd=-min(.$Pc))) %>%
                              tidy %>%
                              dplyr::select(term, estimate) %>% 
                              spread(term, estimate))) %>%
  unnest(model1) %>%
  unnest(data) %>%
  group_by(colony_id)%>%
  summarise(Am=mean(Am), AQY=mean(AQY), Rd=mean(Rd))%>%
  mutate(Ik=Am/AQY)%>%
  mutate(Ic=(Am*Rd)/(AQY*(sqrt(Am^2-Rd^2))))%>%
  write_csv(., "output/pi_curve_pars_nls.csv")

nls_data
```

Plot curve over data points.  
```{r}
augmented <- Data %>% 
  nest(-colony_id) %>% 
  mutate(
    fit = map(data, ~ nls(Pc ~ (Am*((AQY*PAR)/(sqrt(Am^2 + (AQY*PAR)^2)))-Rd), data=., start=list(Am=0.7,  AQY=0.001, Rd=.4))),
    augmented = map(fit, augment),
  ) %>% 
  unnest(augmented)

#all colonies together
augmented %>%
  group_by(colony_id)%>%
  qplot(PAR, Pc, data = ., geom = 'point', colour = colony_id) +
  geom_line(aes(y=.fitted))+
  theme(legend.position="none")



```
