==> Ranalysis/scripts/1_PR_rates_DMB.Rmd <==
---
title: "Photosynthesis and respiration rate calculations"
authors: Hollie Putnam and Jill Ashey
edited by: Danielle Becker-Polinski
date: 20220409
output: html_document
---

# PR data from heatwave timepoint 1 corals
```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## install packages if you dont already have them in your library
if (!require("devtools")) install.packages("devtools")
if (!require("furrr")) install.packages("furrr")
if (!require("future")) install.packages("future")
if (!require("tidyverse")) install.packages("tidyverse")
if (!require("gridExtra")) install.packages("gridExtra")
if (!require("ggpubr")) install.packages("ggpubr")
if (!require("lubridate")) install.packages("lubridate")
if (!require("cowplot")) install.packages("cowplot")
if (!require("LoLinR")) install_github('colin-olito/LoLinR') 
if (!require("ggstatsplot")) install.packages("ggstatsplot")
if ("rTPC" %in% rownames(installed.packages()) == 'FALSE') remotes::install_github("padpadpadpad/rTPC")
if ("nls.multstart" %in% rownames(installed.packages()) == 'FALSE') install.packages('nls.multstart')
if ("broom" %in% rownames(installed.packages()) == 'FALSE') install.packages('broom') 

## load libraries
library(devtools)
library(LoLinR)
library(tidyverse)
library(gridExtra)
library(ggpubr)
library(lubridate)
library(cowplot)
library(ggstatsplot)
library('rTPC')
library('nls.multstart')
library('broom')
library('car')
library('scales')


## libraries for parallel processing
library(future)
library(furrr)
```

## Import data
```{r, warning = FALSE}
path.p <- "data/PR" #the location of all your respirometry files 

# List data files
file.names <- list.files(path = path.p, pattern = "csv$")  # list all csv file names in the folder
file.names <- file.names[!grepl("metadata", file.names)]   # omit metadata from files to be read in as data

# Load PI curve sample metadata (i.e., which corals were in which runs)
sample.info <- read_csv(file = "data/PR/PR_sample_metadata.csv")

# Load PI curve run metadata (i.e., light levels and interval times for each run)
run.info <- read_csv(file = "data/PR/PR_run_metadata.csv")

# Join all coral and run metadata
metadata <- full_join(sample.info, run.info) %>%
  mutate(Date = as_date(as.character(Date), format = "%Y%m%d", tz = "Tahiti"))

# Select only certain columnns
metadata <- metadata %>%
  select(fragment_ID, Run, Treatment, Chamber.Vol.L, Date, Start.time, Stop.time, Light_Value, Temp.Cat)

# Read in all data files
#edit fragment_ID name to remove .csv and change _TP1 to _1
df <- tibble(file.name = file.names) %>%
  mutate(fragment_ID = gsub("_.*", "", file.name)) %>%  # Get fragment_ID from filename and select for info after the first _
  mutate(fragment_ID = gsub(".csv", "", fragment_ID),  #remove .csv from fragment_ID
          info = map(fragment_ID, ~filter(metadata, fragment_ID == .)),           # Get associated sample info
         data0 = map(file.name, ~read_csv(file.path(path.p, .), skip = 1, col_types = cols(.default = "d", Time = "t"))))   # Get associated O2 data

# Select only Time, Value, and Temp columns from O2 data
df <- df %>%
  mutate(data0 = map(data0, ~select(., Time, Value, Temp)))%>%
  mutate(data0 = map(data0, ~(.x %>% filter(complete.cases(.))))) #remove NAs to get rid of artifact line in our data

```

## Use the time breaks in the sample info to link O2 data with light levels
```{r, warning = FALSE}
df <- df %>%
  mutate(intervals = map2(data0, info, function(.x, .y) {
    split(.x, f = cut(as.numeric(.x$Time), breaks = as.numeric(c(.y$Start.time, last(.y$Stop.time))),
                      labels = as.character(.y$Light_Value)))})) %>%
  mutate(data = map(intervals, ~ unnest(tibble(.), .id = "Light_Value")))

## 'data' now contains the O2 data with the corresponding light level as another column
## Example of what 'data' for each sample looks like:
# df$data[[1]]

```

### Thin data
```{r, fig.height = 8, fig.width = 8}
# Set thinning parameter
thin_par <- 20

# Thin data for all samples
df <- df %>%
  mutate(thin_data = map(data, ~ slice(., seq(1, nrow(.), thin_par))))

# Create plots for full dataset and thinned data
df <- df %>%
  mutate(data_plot = map2(data, fragment_ID, ~ ggplot(.x, aes(x = Time, y = Value)) + 
                            facet_wrap(~ as.numeric(Light_Value), scales = "free") +
                            geom_point() +
                            labs(title = .y)),
    thin_data_plot = map2(thin_data, fragment_ID, ~ ggplot(.x, aes(x = Time, y = Value)) + 
                            facet_wrap(~ as.numeric(Light_Value), scales = "free") +
                            geom_point() +
                            labs(title = .y)))

# Example of plots
cowplot::plot_grid(df$data_plot[[1]], df$thin_data_plot[[1]], nrow = 2,
                   labels = c("Example plot: all data", "Example plot: thinned data"))
```

#### The full or thinned data plot for any sample can be accessed like this:
```
df %>%
  filter(fragment_ID == "ACR1_1") %>%
  pull(thin_data_plot)
```

# Fit regressions to each interval for each sample
```{r}
# Define function for fitting LoLinR regressions to be applied to all intervals for all samples
fit_reg <- function(df) {
  rankLocReg(xall = as.numeric(df$Time), yall = df$Value, 
             alpha = 0.2, method = "pc", verbose = FALSE)
}

# Setup for parallel processing
future::plan(multiprocess)

# Map LoLinR function onto all intervals of each sample's thinned dataset
df <- df %>%
  mutate(regs = furrr::future_map(thin_data, function(.) {       # future_map executes function in parallel
    group_by(., Light_Value) %>%
    do(rankLcRg = fit_reg(.))
  }))

## Now 'regs' contains the fitted local regressions for each interval of each sample's thinned dataset

# Define function to pull out and plot regression diagnostics
plot_rankLcRg <- function(fragment_ID, interval_number) {
  df %>%
    filter(fragment_ID == fragment_ID) %>%
    pluck("regs", 1, "rankLcRg", interval_number) %>%
    plot()
}
```

#### The diagnostics for any regression can be plotted like this, specifying a fragment_ID and the number of the light curve interval:
```
plot_rankLcRg("BLK032-38", 1)
```

### Extract slope of best regression for each interval for each sample
```{r}
df.out <- df %>% 
  unnest(regs) %>%
  mutate(micromol.L.s = map_dbl(rankLcRg, ~ pluck(., "allRegs", "b1", 1)))
```

# Adjust by chamber volume and normalize to surface area
```{r}
### Merge rates with sample info
pr <- left_join(
  select(df.out, fragment_ID, Light_Value, micromol.L.s),
  distinct(metadata, fragment_ID, Treatment, Run, Chamber.Vol.L, Date)
)

### Correct for chamber volume and blanks
pr <- pr %>%
  mutate(micromol.s = micromol.L.s * Chamber.Vol.L)

# Get blank values -- average for each run and light value in case multiple blanks
blanks <- pr %>%
  filter(grepl("BLK", fragment_ID)) %>%
  group_by(Run, Light_Value) %>%
  summarise(micromol.s.blank = mean(micromol.s))

# Join blank values with rest of data and subtract values from samples for same run and light value
pr <- left_join(pr, blanks) %>%
  mutate(micromol.s.adj = micromol.s - micromol.s.blank) %>%
  # After correcting for blank values, remove blanks from data
  filter(!grepl("BLK", fragment_ID))


# Import surface area data
sa <- read_csv(file = "data/surface.area.calc.csv")

# Join surface area with rest of data
pr <- left_join(pr, select(sa, fragment_ID, surface.area.cm2))

# Normalize rates by surface area
pr <- pr %>%
  mutate(micromol.cm2.s = micromol.s.adj / surface.area.cm2,
         micromol.cm2.h = micromol.cm2.s * 3600)

Photo <- pr %>%
  filter(Light_Value ==620)%>% 
  separate(fragment_ID, c("Species","Number","Temp.Cat"), sep = "-")

Resp <- pr %>%
  filter(Light_Value ==0)%>% 
  separate(fragment_ID, c("Species","Number","Temp.Cat"), sep = "-")

# Import morphology data
morph<- read_csv(file = "data/morphology.csv")
morph$Number <- as.character(morph$Number)

# Join morph with rest of data
Photo <- left_join(Photo, morph)
Resp <- left_join(Resp, morph)

Photo <- na.omit(Photo)
Resp <- na.omit(Resp)
```

# Plot rates vs. temperature for each sample
```{r, fig.height=15, fig.width = 10}
# Photo %>% group_by(fragment_ID) %>%
# ggplot(aes(x = Temp.Cat, y = micromol.cm2.h, color=fragment_ID, group =fragment_ID )) +
#   geom_point(aes(size=6, group=fragment_ID))+
#   geom_line(aes(group=fragment_ID))

#pdf("output/Photo_TPC1.pdf", width=6, height=6)
Photo %>% 
ggplot(aes(x=Temp.Cat, y=micromol.cm2.h, color=Morphology, group=Number)) +
  geom_point(size=3)+
  geom_path()
#dev.off()

#pdf("output/Resp_TPC1.pdf", width=6, height=6)
Resp %>% 
ggplot(aes(x=Temp.Cat, y=-micromol.cm2.h, color=Morphology, group=Number)) +
  geom_point(size=3)+
  geom_path()
#dev.off()

```

```{r}
# Visualize data
pdf("output/Resp_TPC.pdf", width=6, height=6)
r_plot<-Resp %>%
    ggplot(., aes(x = Temp.Cat, y = -micromol.cm2.h, colour=Morphology)) +
    geom_point(aes(fill=Morphology, group=Number), pch = 21, size=2, alpha=0.3) + 
    xlab("Temperature") + 
    scale_fill_manual(name="Morphology", values=c("orange","black","cyan"))+
    scale_color_manual(name="Morphology", values=c("orange","black","cyan"))+
    ylab(expression(bold(paste("R (µmol ", O[2], " cm"^-2, "min"^-1, ")")))) +
    theme_classic() + 
    theme(axis.title=element_text(face="bold", size=16),
      axis.text=element_text(size=12, color="black"), 
      legend.title=element_text(face="bold", size=14), 
      legend.text=element_text(size=12)); r_plot
dev.off()

pdf("output/Photo_TPC.pdf", width=6, height=6)
r_plot<-Photo %>%
    ggplot(., aes(x = Temp.Cat, y = micromol.cm2.h, colour=Morphology)) +
    geom_point(aes(fill=Morphology, group=Number), pch = 21, size=2, alpha=0.3) + 
    xlab("Temperature") + 
    scale_fill_manual(name="Morphology", values=c("orange","black","cyan"))+
    scale_color_manual(name="Morphology", values=c("orange","black","cyan"))+
    ylab(expression(bold(paste("R (µmol ", O[2], " cm"^-2, "min"^-1, ")")))) +
    theme_classic() + 
    theme(axis.title=element_text(face="bold", size=16),
      axis.text=element_text(size=12, color="black"), 
      legend.title=element_text(face="bold", size=14), 
      legend.text=element_text(size=12)); r_plot
dev.off()
```

PHOTOSYNTHESIS View and remove outliers.   
```{r, warning=FALSE, message=FALSE}
#identify outliers by Temperature and Treatment groups
outlier.plot <- ggbetweenstats(Photo,Morphology, micromol.cm2.h, outlier.tagging = TRUE)
outlier.plot

#set quantile values
q <- c(0.25, 0.75)

# calculate quantile values by Temperature and Treatment groups
Quants <- Photo %>%
  group_by(Morphology, Temp.Cat) %>%
  summarize(quant25 = quantile(micromol.cm2.h, probs = q[1]),
            quant75 = quantile(micromol.cm2.h, probs = q[2]),
            IQRbyGroup=IQR(micromol.cm2.h))

# add a group name by pasting Temperature and Treatment
Quants$group <-paste0(Quants$Morphology,"_", Quants$Temp.Cat)

#Calculate Quantile upper and lower ranges 
Quants$upper <-  Quants$quant75+1.5*Quants$IQRbyGroup # Upper Range  
Quants$lower <- Quants$quant25-1.5*Quants$IQRbyGroup # Lower Range

#join outlier cutoffs with rate data
Photodata <- left_join(Photo, Quants)

#remove outliers from rates
Photodata <- Photodata %>%
  filter(micromol.cm2.h < upper) %>%
  filter(micromol.cm2.h > lower) #%>%
  #filter(rate < 0.125)

# Visualize data following outlier removal
P_plot<-Photodata %>%
    ggplot(., aes(x = Temp.Cat, y = micromol.cm2.h, colour=Morphology)) +
    geom_point(aes(fill=Morphology, group=Number), pch = 21, size=2, alpha=0.3) + 
    xlab("Temperature") + 
    scale_fill_manual(name="Morphology", values=c("orange","black","cyan"))+
    scale_color_manual(name="Morphology", values=c("orange","black","cyan"))+
    ylab(expression(bold(paste("R (µmol ", O[2], " cm"^-2, "min"^-1, ")")))) +
    theme_classic() + 
    theme(legend.position="none",
      axis.title=element_text(face="bold", size=16),
      axis.text=element_text(size=12, color="black"), 
      legend.title=element_text(face="bold", size=14), 
      legend.text=element_text(size=12)); P_plot

```

RESPIRATION View and remove outliers.   
```{r, warning=FALSE, message=FALSE}
#identify outliers by Temperature and Treatment groups
outlier.plot <- ggbetweenstats(Resp,Morphology, micromol.cm2.h, outlier.tagging = TRUE)
outlier.plot

#set quantile values
q <- c(0.25, 0.75)

# calculate quantile values by Temperature and Treatment groups
Quants <- Resp %>%
  group_by(Morphology, Temp.Cat) %>%
  summarize(quant25 = quantile(micromol.cm2.h, probs = q[1]),
            quant75 = quantile(micromol.cm2.h, probs = q[2]),
            IQRbyGroup=IQR(micromol.cm2.h))

# add a group name by pasting Temperature and Treatment
Quants$group <-paste0(Quants$Morphology,"_", Quants$Temp.Cat)

#Calculate Quantile upper and lower ranges 
Quants$upper <-  Quants$quant75+1.5*Quants$IQRbyGroup # Upper Range  
Quants$lower <- Quants$quant25-1.5*Quants$IQRbyGroup # Lower Range

#join outlier cutoffs with rate data
Respdata <- left_join(Resp, Quants)

#remove outliers from rates
Respdata <- Respdata %>%
  filter(micromol.cm2.h < upper) %>%
  filter(micromol.cm2.h > lower) #%>%
  #filter(rate < 0.125)

#Respdata$micromol.cm2.h <- -Respdata$micromol.cm2.h

# Visualize data following outlier removal
R_plot<-Respdata %>%
    ggplot(., aes(x = Temp.Cat, y = -micromol.cm2.h, colour=Morphology)) +
    geom_point(aes(fill=Morphology, group=Number), pch = 21, size=2, alpha=0.3) + 
    xlab("Temperature") + 
    scale_fill_manual(name="Morphology", values=c("orange","black","cyan"))+
    scale_color_manual(name="Morphology", values=c("orange","black","cyan"))+
    ylab(expression(bold(paste("R (µmol ", O[2], " cm"^-2, "min"^-1, ")")))) +
    theme_classic() + 
    theme(legend.position="none",
      axis.title=element_text(face="bold", size=16),
      axis.text=element_text(size=12, color="black"), 
      legend.title=element_text(face="bold", size=14), 
      legend.text=element_text(size=12)); R_plot

```

TPC fitting 
Padifeld et al **rTPC and nls.multstart: A new pipeline to fit thermal performance curves in r**  
https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13585  

Sharpe Schoolfield 1981 model
Schoolfield, R. M., Sharpe, P. J. H., & Magnuson, C. E. (1981). Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. Journal of theoretical biology, 88(4), 719-731. https://doi.org/10.1016/0022-5193(81)90246-0

```{r}
#Photosynthesis
Photodata$temp <- as.numeric(Photodata$Temp.Cat)
#Photodata <- Photodata %>% mutate(micromol.cm2.h = replace(micromol.cm2.h, which(micromol.cm2.h<0), 0))
Photodata$micromol.cm2.h <- Photodata$micromol.cm2.h+1

# choose model
get_model_names()
#sharpeschoolhigh_1981

# get start vals
start_vals <- get_start_vals(Photodata$temp,Photodata$micromol.cm2.h, model_name = 'sharpeschoolhigh_1981')

# get limits
low_lims <- get_lower_lims(Photodata$temp,Photodata$micromol.cm2.h, model_name = 'sharpeschoolhigh_1981')
upper_lims <- get_upper_lims(Photodata$temp,Photodata$micromol.cm2.h, model_name = 'sharpeschoolhigh_1981')

#view values
start_vals
#start_vals[2] <- 0.1
#start_vals[3] <- 0.1
#start_vals
low_lims
upper_lims

#GRANDIS
# grandis CURVE FIT
d.grandis <- Photodata %>% 
  filter(Morphology=="grandis")

#fit 
grandis.fit <- nls_multstart(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.grandis,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

grandis.fit

#generate the predicted data
grandis_new_data <- data.frame(temp = seq(min(d.grandis$temp), max(d.grandis$temp), 0.5))
grandis.preds <- augment(grandis.fit, newdata = grandis_new_data)

#calculate TPC parameters
grandis.TCP.res <- calc_params(grandis.fit) %>%
  mutate_all(round, 2)   # round 

grandis.TCP.res 

### Bootstrapping ambient curve fit    
# refit model using nlsLM
grandis.fit_nlsLM <- minpack.lm::nlsLM(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.grandis,
                        start = coef(grandis.fit),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.grandis)))

# bootstrap using case resampling
grandis.boot1 <- Boot(grandis.fit_nlsLM, method = 'case')

# look at the data
head(grandis.boot1$t)


# create predictions of each bootstrapped model
grandis.boot1_preds <- grandis.boot1$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.grandis$temp), max(d.grandis$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
grandis.boot1_conf_preds <- group_by(grandis.boot1_preds, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
grandis.CI.plot <- ggplot() +
  geom_line(aes(temp, .fitted), grandis.preds, col = 'orange') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds, fill = 'orange', alpha = 0.3) +
  geom_point(aes(temp, micromol.cm2.h), d.grandis, size = 2, alpha = 0.5,col = 'orange') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Rate µmol O2/cm2/h')
grandis.CI.plot 

```

```{r}

#MEANDRINA
# meandrina CURVE FIT
d.meandrina <- Photodata %>% 
  filter(Morphology=="meandrina")

#fit 
meandrina.fit <- nls_multstart(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.meandrina,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

meandrina.fit

#generate the predicted data
meandrina_new_data <- data.frame(temp = seq(min(d.meandrina$temp), max(d.meandrina$temp), 0.5))
meandrina.preds <- augment(meandrina.fit, newdata = meandrina_new_data)

#calculate TPC parameters
meandrina.TCP.res <- calc_params(meandrina.fit) %>%
  mutate_all(round, 2)   # round 

meandrina.TCP.res 

# refit model using nlsLM
meandrina.fit_nlsLM <- minpack.lm::nlsLM(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.meandrina,
                        start = coef(meandrina.fit),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.meandrina)))

# bootstrap using case resampling
meandrina.boot1 <- Boot(meandrina.fit_nlsLM, method = 'case')

# look at the data
head(meandrina.boot1$t)


# create predictions of each bootstrapped model
meandrina.boot1_preds <- meandrina.boot1$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.meandrina$temp), max(d.meandrina$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
meandrina.boot1_conf_preds <- group_by(meandrina.boot1_preds, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
meandrina.CI.plot <- ggplot() +
  geom_line(aes(temp, .fitted), meandrina.preds, col = 'black') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), meandrina.boot1_conf_preds, fill = 'black', alpha = 0.3) +
  geom_point(aes(temp, micromol.cm2.h), d.meandrina, size = 2, alpha = 0.5,col = 'black') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Rate (nmol O2/cm2/h')
meandrina.CI.plot 

```


```{r}

#VERRUCOSA
# verrucosa CURVE FIT
d.verrucosa <- Photodata %>% 
  filter(Morphology=="verrucosa")

#fit 
verrucosa.fit <- nls_multstart(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.verrucosa,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

verrucosa.fit

#generate the predicted data
verrucosa_new_data <- data.frame(temp = seq(min(d.verrucosa$temp), max(d.verrucosa$temp), 0.5))
verrucosa.preds <- augment(verrucosa.fit, newdata = verrucosa_new_data)

#calculate TPC parameters
verrucosa.TCP.res <- calc_params(verrucosa.fit) %>%
  mutate_all(round, 2)   # round 

verrucosa.TCP.res 

# refit model using nlsLM
verrucosa.fit_nlsLM <- minpack.lm::nlsLM(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.verrucosa,
                        start = coef(verrucosa.fit),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.verrucosa)))

# bootstrap using case resampling
verrucosa.boot1 <- Boot(verrucosa.fit_nlsLM, method = 'case')

# look at the data
head(verrucosa.boot1$t)


# create predictions of each bootstrapped model
verrucosa.boot1_preds <- verrucosa.boot1$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.verrucosa$temp), max(d.verrucosa$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
verrucosa.boot1_conf_preds <- group_by(verrucosa.boot1_preds, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
verrucosa.CI.plot <- ggplot() +
  geom_line(aes(temp, .fitted), verrucosa.preds, col = 'cyan') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds, fill = 'cyan', alpha = 0.3) +
  geom_point(aes(temp, micromol.cm2.h), d.verrucosa, size = 2, alpha = 0.5,col = 'cyan') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Rate (nmol O2/cm2/h')
verrucosa.CI.plot

```

```{r}
 #set plot colors
cols <- c("grandis"="orange", "meandrina"="black", "verrucosa"="cyan")
  
# plot data and model fit
TPC.plot <- ggplot(data=Photodata, aes(x=temp)) +
   geom_point(aes(temp, micromol.cm2.h, color="grandis"), d.grandis, size = 2, alpha = 0.5) +
   geom_point(aes(temp, micromol.cm2.h, color="meandrina"), d.meandrina, size = 2, alpha = 0.5) +
     geom_point(aes(temp, micromol.cm2.h, color="verrucosa"), d.verrucosa, size = 2, alpha = 0.5) +
   geom_line(aes(temp, .fitted), grandis.preds, col = 'orange', size=2) +
   geom_line(aes(temp, .fitted), meandrina.preds, col = "black", size=2) +
   geom_line(aes(temp, .fitted), verrucosa.preds, col = "cyan", size=2) +
   geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds, fill = "orange", alpha = 0.3) +
   geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), meandrina.boot1_conf_preds, fill = 'black', alpha = 0.3) +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds, fill = 'cyan', alpha = 0.3) +
   xlim(21,37)+
   scale_x_continuous(breaks=c(22,24,26,28,30,32,34,36))+
   theme_bw(base_size = 12) +
   scale_colour_manual(name="Morphology",values=cols)+
   theme(legend.position = c(0.55, 0.2),
         panel.border = element_blank(), panel.grid.major = element_blank(),
         panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))+
   labs(x = 'Temperature (ºC)',
        y = expression("Rate"~µmol~O[2] ~cm^{-2}~h^{-1}))

TPC.plot 

ggsave("output/TPC_SharpSchool_Photosynthesis.pdf", TPC.plot, dpi=300, w=8, h=8, units="in")
```
```{r}
broom::tidy(grandis.fit_nlsLM)
broom::tidy(meandrina.fit_nlsLM)
broom::tidy(verrucosa.fit_nlsLM)

#GRANDIS
#calculate all the TPC parameters
grandis.extra_params <- calc_params(grandis.fit_nlsLM) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
grandis.ci_extra_params <- Boot(grandis.fit_nlsLM, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(grandis.fit_nlsLM)), R = 200, method = 'case') %>%
  confint(., method = 'bca') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')

#join the parameters and CIs  
grandis.ci_extra_params <- left_join(grandis.ci_extra_params, grandis.extra_params)
grandis.ci_extra_params$Treatment <- "P. grandis"

#MEANDRINA
#calculate all the TPC parameters
meandrina.extra_params <- calc_params(meandrina.fit_nlsLM) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
meandrina.ci_extra_params <- Boot(meandrina.fit_nlsLM, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(meandrina.fit_nlsLM)), R = 200, method = 'case') %>%
  confint(., method = 'bca') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')
  
#join the parameters and CIs  
meandrina.ci_extra_params <- left_join(meandrina.ci_extra_params, meandrina.extra_params)
meandrina.ci_extra_params$Treatment <- "P. meandrina"

#VERRUCOSA
#calculate all the TPC parameters
verrucosa.extra_params <- calc_params(verrucosa.fit_nlsLM) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
verrucosa.ci_extra_params <- Boot(verrucosa.fit_nlsLM, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(verrucosa.fit_nlsLM)), R = 200, method = 'case') %>%
  confint(., method = 'bca') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')
  
#join the parameters and CIs  
verrucosa.ci_extra_params <- left_join(verrucosa.ci_extra_params, verrucosa.extra_params)
verrucosa.ci_extra_params$Treatment <- "P. verrucosa"

#Join Ambient and High estimates and CIs
All_params <- rbind(grandis.ci_extra_params, meandrina.ci_extra_params, verrucosa.ci_extra_params)
All_params <- All_params %>% 
 mutate_if(is.numeric, round, 2)

#Plot all of the estimates
estimate.plots <- ggplot(All_params, aes(Treatment, estimate, color=Treatment)) +
  geom_point(size = 2) +
  scale_color_manual(name="Treatment", values=c("orange","black", "cyan"))+
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  facet_wrap(~param, scales = 'free_y') +
  scale_x_discrete('')

estimate.plots

#filter to only the most relavent and well characterized parameters
All_params <- All_params %>% 
  filter(!param=="ctmin") %>%
  filter(!param=="ctmax") %>%
  filter(!param=="skewness") %>%
  filter(!param=="thermal_safety_margin") %>%
  filter(!param=="thermal_tolerance") %>%
  filter(!param=="q10") %>%
  filter(!param=="breadth")
  
#view estimate plots
estimate.plots <- ggplot(All_params, aes(Treatment, estimate, color=Treatment)) +
  geom_point(size = 2) +
  scale_color_manual(name="Treatment", values=c("black","cyan", "orange"))+
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  labs(y = NULL)+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        legend.position = "none",
        strip.background = element_blank(), 
        strip.placement = "outside") +
  facet_wrap(~param, scales = 'free_y', nrow=1)+
             #labeller = as_labeller(c(e = "e (Energy)", eh = " eh (Energy)", rmax= "Rmax (~nmol~O[2] ~larva^{-1}~min^{-1})",topt="Topt (Temperature °C)")), strip.position = "left") +
  scale_x_discrete('')

estimate.plots

ggsave("output/TPC_estimates_SharpSchool_Photosynthesis.pdf", estimate.plots, dpi=300, w=6, h=2, units="in")
``` 

```{r}
#Respiration
Respdata$temp <- as.numeric(Respdata$Temp.Cat)
Respdata$micromol.cm2.h <- -Respdata$micromol.cm2.h+1

# choose model
get_model_names()
#sharpeschoolhigh_1981

# get start vals
start_vals <- get_start_vals(Respdata$temp,Respdata$micromol.cm2.h, model_name = 'sharpeschoolhigh_1981')

# get limits
low_lims <- get_lower_lims(Respdata$temp,Respdata$micromol.cm2.h, model_name = 'sharpeschoolhigh_1981')
upper_lims <- get_upper_lims(Respdata$temp,Respdata$micromol.cm2.h, model_name = 'sharpeschoolhigh_1981')

#view values
start_vals
start_vals[3] <- 0.1
start_vals
low_lims
upper_lims

#GRANDIS
# grandis CURVE FIT
d.grandis <- Respdata %>% 
  filter(Morphology=="grandis")

#d.grandis$micromol.cm2.h <- log10(d.grandis$micromol.cm2.h) 


#fit 
grandis.fit <- nls_multstart(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.grandis,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

grandis.fit

#generate the predicted data
grandis_new_data <- data.frame(temp = seq(min(d.grandis$temp), max(d.grandis$temp), 0.5))
grandis.preds <- augment(grandis.fit, newdata = grandis_new_data)

#calculate TPC parameters
grandis.TCP.res <- calc_params(grandis.fit) %>%
  mutate_all(round, 2)   # round 

grandis.TCP.res 

### Bootstrapping ambient curve fit    
# refit model using nlsLM
grandis.fit_nlsLM <- minpack.lm::nlsLM(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.grandis,
                        start = coef(grandis.fit),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.grandis)))

# bootstrap using case resampling
grandis.boot1 <- Boot(grandis.fit_nlsLM, method = 'case')

# look at the data
head(grandis.boot1$t)


# create predictions of each bootstrapped model
grandis.boot1_preds <- grandis.boot1$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.grandis$temp), max(d.grandis$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
grandis.boot1_conf_preds <- group_by(grandis.boot1_preds, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
grandis.CI.plot <- ggplot() +
  geom_line(aes(temp, .fitted), grandis.preds, col = 'orange') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds, fill = 'orange', alpha = 0.3) +
  geom_point(aes(temp, micromol.cm2.h), d.grandis, size = 2, alpha = 0.5,col = 'orange') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Rate (µmol O2/cm2/h')
grandis.CI.plot 

```

```{r}

#MEANDRINA
# meandrina CURVE FIT
d.meandrina <- Respdata %>% 
  filter(Morphology=="meandrina")

#fit 
meandrina.fit <- nls_multstart(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.meandrina,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

meandrina.fit

#generate the predicted data
meandrina_new_data <- data.frame(temp = seq(min(d.meandrina$temp), max(d.meandrina$temp), 0.5))
meandrina.preds <- augment(meandrina.fit, newdata = meandrina_new_data)

#calculate TPC parameters
meandrina.TCP.res <- calc_params(meandrina.fit) %>%
  mutate_all(round, 2)   # round 

meandrina.TCP.res 

# refit model using nlsLM
meandrina.fit_nlsLM <- minpack.lm::nlsLM(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.meandrina,
                        start = coef(meandrina.fit),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.meandrina)))

# bootstrap using case resampling
meandrina.boot1 <- Boot(meandrina.fit_nlsLM, method = 'case')

# look at the data
head(meandrina.boot1$t)


# create predictions of each bootstrapped model
meandrina.boot1_preds <- meandrina.boot1$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.meandrina$temp), max(d.meandrina$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
meandrina.boot1_conf_preds <- group_by(meandrina.boot1_preds, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
meandrina.CI.plot <- ggplot() +
  geom_line(aes(temp, .fitted), meandrina.preds, col = 'black') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), meandrina.boot1_conf_preds, fill = 'black', alpha = 0.3) +
  geom_point(aes(temp, micromol.cm2.h), d.meandrina, size = 2, alpha = 0.5,col = 'black') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Rate (nmol O2/cm2/h')
meandrina.CI.plot 

```


```{r}

#VERRUCOSA
# verrucosa CURVE FIT
d.verrucosa <- Respdata %>% 
  filter(Morphology=="verrucosa")

#fit 
verrucosa.fit <- nls_multstart(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.verrucosa,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

verrucosa.fit

#generate the predicted data
verrucosa_new_data <- data.frame(temp = seq(min(d.verrucosa$temp), max(d.verrucosa$temp), 0.5))
verrucosa.preds <- augment(verrucosa.fit, newdata = verrucosa_new_data)

#calculate TPC parameters
verrucosa.TCP.res <- calc_params(verrucosa.fit) %>%
  mutate_all(round, 2)   # round 

verrucosa.TCP.res 

# refit model using nlsLM
verrucosa.fit_nlsLM <- minpack.lm::nlsLM(micromol.cm2.h~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.verrucosa,
                        start = coef(verrucosa.fit),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.verrucosa)))

# bootstrap using case resampling
verrucosa.boot1 <- Boot(verrucosa.fit_nlsLM, method = 'case')

# look at the data
head(verrucosa.boot1$t)


# create predictions of each bootstrapped model
verrucosa.boot1_preds <- verrucosa.boot1$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.verrucosa$temp), max(d.verrucosa$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
verrucosa.boot1_conf_preds <- group_by(verrucosa.boot1_preds, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
verrucosa.CI.plot <- ggplot() +
  geom_line(aes(temp, .fitted), verrucosa.preds, col = 'cyan') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds, fill = 'cyan', alpha = 0.3) +
  geom_point(aes(temp, micromol.cm2.h), d.verrucosa, size = 2, alpha = 0.5,col = 'cyan') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Rate (nmol O2/cm2/h')
verrucosa.CI.plot

```



```{r}
 #set plot colors
cols <- c("grandis"="orange", "meandrina"="black", "verrucosa"="cyan")
  
# plot data and model fit
TPC.plot <- ggplot(data=Respdata, aes(x=temp)) +
   geom_point(aes(temp, micromol.cm2.h, color="grandis"), d.grandis, size = 2, alpha = 0.5) +
   geom_point(aes(temp, micromol.cm2.h, color="meandrina"), d.meandrina, size = 2, alpha = 0.5) +
     geom_point(aes(temp, micromol.cm2.h, color="verrucosa"), d.verrucosa, size = 2, alpha = 0.5) +
   geom_line(aes(temp, .fitted), grandis.preds, col = 'orange', size=2) +
   geom_line(aes(temp, .fitted), meandrina.preds, col = "black", size=2) +
   geom_line(aes(temp, .fitted), verrucosa.preds, col = "cyan", size=2) +
   geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds, fill = "orange", alpha = 0.3) +
   geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), meandrina.boot1_conf_preds, fill = 'black', alpha = 0.3) +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds, fill = 'cyan', alpha = 0.3) +
   xlim(21,37)+
   scale_x_continuous(breaks=c(22,24,26,28,30,32,34,36))+
   theme_bw(base_size = 12) +
   scale_colour_manual(name="Morphology",values=cols)+
   theme(legend.position = c(0.55, 0.2),
         panel.border = element_blank(), panel.grid.major = element_blank(),
         panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))+
   labs(x = 'Temperature (ºC)',
        y = expression("Rate"~µmol~O[2] ~cm^{-2}~h^{-1}))

TPC.plot 

ggsave("output/TPC_SharpSchool_Respiration.pdf", TPC.plot, dpi=300, w=8, h=8, units="in")
### Confidence intervals of TPC parameters   
```



```{r}
broom::tidy(grandis.fit_nlsLM)
broom::tidy(meandrina.fit_nlsLM)
broom::tidy(verrucosa.fit_nlsLM)

#GRANDIS
#calculate all the TPC parameters
grandis.extra_params <- calc_params(grandis.fit_nlsLM) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
grandis.ci_extra_params <- Boot(grandis.fit_nlsLM, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(grandis.fit_nlsLM)), R = 200, method = 'case') %>%
  confint(., method = 'bca') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')

#join the parameters and CIs  
grandis.ci_extra_params <- left_join(grandis.ci_extra_params, grandis.extra_params)
grandis.ci_extra_params$Treatment <- "P. grandis"

#MEANDRINA
#calculate all the TPC parameters
meandrina.extra_params <- calc_params(meandrina.fit_nlsLM) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
meandrina.ci_extra_params <- Boot(meandrina.fit_nlsLM, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(meandrina.fit_nlsLM)), R = 200, method = 'case') %>%
  confint(., method = 'bca') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')
  
#join the parameters and CIs  
meandrina.ci_extra_params <- left_join(meandrina.ci_extra_params, meandrina.extra_params)
meandrina.ci_extra_params$Treatment <- "P. meandrina"

#VERRUCOSA
#calculate all the TPC parameters
verrucosa.extra_params <- calc_params(verrucosa.fit_nlsLM) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
verrucosa.ci_extra_params <- Boot(verrucosa.fit_nlsLM, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(verrucosa.fit_nlsLM)), R = 200, method = 'case') %>%
  confint(., method = 'bca') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')
  
#join the parameters and CIs  
verrucosa.ci_extra_params <- left_join(verrucosa.ci_extra_params, verrucosa.extra_params)
verrucosa.ci_extra_params$Treatment <- "P. verrucosa"

#Join Ambient and High estimates and CIs
All_params <- rbind(grandis.ci_extra_params, meandrina.ci_extra_params, verrucosa.ci_extra_params)
All_params <- All_params %>% 
 mutate_if(is.numeric, round, 2)

#Plot all of the estimates
estimate.plots <- ggplot(All_params, aes(Treatment, estimate, color=Treatment)) +
  geom_point(size = 2) +
  scale_color_manual(name="Treatment", values=c("black","cyan", "orange"))+
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  facet_wrap(~param, scales = 'free_y') +
  scale_x_discrete('')

estimate.plots

#filter to only the most relavent and well characterized parameters
All_params <- All_params %>% 
  filter(!param=="ctmin") %>%
  filter(!param=="ctmax") %>%
  filter(!param=="skewness") %>%
  filter(!param=="thermal_safety_margin") %>%
  filter(!param=="thermal_tolerance") %>%
  filter(!param=="q10") %>%
  filter(!param=="breadth")
  
#view estimate plots
estimate.plots <- ggplot(All_params, aes(Treatment, estimate, color=Treatment)) +
  geom_point(size = 2) +
  scale_color_manual(name="Treatment", values=c("black","cyan", "orange"))+
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  labs(y = NULL)+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        legend.position = "none",
        strip.background = element_blank(), 
        strip.placement = "outside") +
  facet_wrap(~param, scales = 'free_y', nrow=1)+
             #labeller = as_labeller(c(e = "e (Energy)", eh = " eh (Energy)", rmax= "Rmax (~nmol~O[2] ~larva^{-1}~min^{-1})",topt="Topt (Temperature °C)")), strip.position = "left") +
  scale_x_discrete('')

estimate.plots

ggsave("output/TPC_estimates_SharpSchool_Respiration.pdf", estimate.plots, dpi=300, w=6, h=2, units="in")
```
