==> Ranalysis/scripts/3_May_Dec_Photo_rate_TPC_fitting.Rmd <==
---
title: "Photosynthesis TPC fitting"
authors: Hollie Putnam 
edited by: Hollie Putnam 
date: 20220829
output: html_document
---

# PR data from heatwave timepoint 1 corals
```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## install packages if you dont already have them in your library
if (!require("devtools")) install.packages("devtools")
if (!require("furrr")) install.packages("furrr")
if (!require("future")) install.packages("future")
if (!require("tidyverse")) install.packages("tidyverse")
if (!require("gridExtra")) install.packages("gridExtra")
if (!require("ggpubr")) install.packages("ggpubr")
if (!require("lubridate")) install.packages("lubridate")
if (!require("cowplot")) install.packages("cowplot")
if (!require("ggstatsplot")) install.packages("ggstatsplot")
if ("rTPC" %in% rownames(installed.packages()) == 'FALSE') remotes::install_github("padpadpadpad/rTPC")
if ("nls.multstart" %in% rownames(installed.packages()) == 'FALSE') install.packages('nls.multstart')
if ("broom" %in% rownames(installed.packages()) == 'FALSE') install.packages('broom') 

## load libraries
library('devtools')
library('tidyverse')
library('gridExtra')
library('ggpubr')
library('lubridate')
library('cowplot')
library('ggstatsplot')
library('rTPC')
library('nls.multstart')
library('broom')
library('car')
library('scales')
library('future') ## libraries for parallel processing
library('furrr') ## libraries for parallel processing
```

## Import data
```{r, warning = FALSE}
Photodata_May <- read.csv("data/May_Photo_extracted_rates.csv")
Photodata_Dec <- read.csv("data/Dec_Photo_extracted_rates.csv")
Photodata <- rbind(Photodata_May,Photodata_Dec)

```

TPC fitting 
Padifeld et al **rTPC and nls.multstart: A new pipeline to fit thermal performance curves in r**  
https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13585  

Sharpe Schoolfield 1981 model
Schoolfield, R. M., Sharpe, P. J. H., & Magnuson, C. E. (1981). Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. Journal of theoretical biology, 88(4), 719-731. https://doi.org/10.1016/0022-5193(81)90246-0

```{r}
#Photosynthesis
Photodata$temp <- as.numeric(Photodata$Temp.Cat)
Photodata$micromol.cm2.h <- replace(Photodata$micromol.cm2.h, Photodata$micromol.cm2.h<0,0)
Photodata$transformed.rate <- Photodata$micromol.cm2.h
#Photodata$transformed.rate <- log10(Photodata$micromol.cm2.h+1)

# choose model
get_model_names()
#sharpeschoolhigh_1981

# get start vals
start_vals <- get_start_vals(Photodata$temp,Photodata$transformed.rate, model_name = 'sharpeschoolhigh_1981')

# get limits
low_lims <- get_lower_lims(Photodata$temp,Photodata$transformed.rate, model_name = 'sharpeschoolhigh_1981')
upper_lims <- get_upper_lims(Photodata$temp,Photodata$transformed.rate, model_name = 'sharpeschoolhigh_1981')

#view values
start_vals
start_vals <- start_vals %>% replace(is.na(.), 0)
start_vals
low_lims
upper_lims
```

#GRANDIS May
```{r}
# grandis CURVE FIT
# Haplotype_1a
d.grandis.may <- Photodata %>% 
  filter(mtORF=="Haplotype_1a")%>% 
filter(Season=="May")

#fit 
grandis.fit.may <- nls_multstart(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.grandis.may,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

grandis.fit.may

#generate the predicted data
grandis_new_data.may <- data.frame(temp = seq(min(d.grandis.may$temp), max(d.grandis.may$temp), 0.5))
grandis.preds.may <- augment(grandis.fit.may, newdata = grandis_new_data.may)

#calculate TPC parameters
grandis.TCP.res.may <- calc_params(grandis.fit.may) %>%
  mutate_all(round, 2)   # round 

grandis.TCP.res.may 

### Bootstrapping ambient curve fit    
# refit model using nlsLM
grandis.fit_nlsLM.may <- minpack.lm::nlsLM(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.grandis.may,
                        start = coef(grandis.fit.may),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.grandis.may)))

# bootstrap using case resampling
grandis.boot1.may <- Boot(grandis.fit_nlsLM.may, method = 'case')

# look at the data
head(grandis.boot1.may$t)


# create predictions of each bootstrapped model
grandis.boot1_preds.may <- grandis.boot1.may$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.grandis.may$temp), max(d.grandis.may$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
grandis.boot1_conf_preds.may <- group_by(grandis.boot1_preds.may, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
grandis.CI.plot.May <- ggplot() +
  geom_line(aes(temp, .fitted), grandis.preds.may, col = 'orange') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds.may, fill = 'orange', alpha = 0.3) +
  geom_point(aes(temp, transformed.rate), d.grandis.may, size = 2, alpha = 0.5,col = 'orange') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Log 10 +1 Rate (µmol O2/cm2/h)')
grandis.CI.plot.May 

```
#GRANDIS December
```{r}
# grandis CURVE FIT
# Haplotype_1a
d.grandis.dec <- Photodata %>% 
  filter(mtORF=="Haplotype_1a")%>% 
filter(Season=="Dec")

#fit 
grandis.fit.dec <- nls_multstart(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.grandis.dec,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

grandis.fit.dec

#generate the predicted data
grandis_new_data.dec <- data.frame(temp = seq(min(d.grandis.dec$temp), max(d.grandis.dec$temp), 0.5))
grandis.preds.dec <- augment(grandis.fit.dec, newdata = grandis_new_data.dec)

#calculate TPC parameters
grandis.TCP.res.dec <- calc_params(grandis.fit.dec) %>%
  mutate_all(round, 2)   # round 

grandis.TCP.res.dec 

### Bootstrapping ambient curve fit    
# refit model using nlsLM
grandis.fit_nlsLM.dec <- minpack.lm::nlsLM(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.grandis.dec,
                        start = coef(grandis.fit.dec),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.grandis.dec)))

# bootstrap using case resampling
grandis.boot1.dec <- Boot(grandis.fit_nlsLM.dec, method = 'case')

# look at the data
head(grandis.boot1.dec$t)


# create predictions of each bootstrapped model
grandis.boot1_preds.dec <- grandis.boot1.dec$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.grandis.dec$temp), max(d.grandis.dec$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
grandis.boot1_conf_preds.dec <- group_by(grandis.boot1_preds.dec, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
grandis.CI.plot.dec <- ggplot() +
  geom_line(aes(temp, .fitted), grandis.preds.dec, col = 'orange') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds.dec, fill = 'orange', alpha = 0.3) +
  geom_point(aes(temp, transformed.rate), d.grandis.dec, size = 2, alpha = 0.5,col = 'orange') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Log 10 +1 Rate (µmol O2/cm2/h)')
grandis.CI.plot.dec 

```

#VERRUCOSA May
```{r}
# verrucosa CURVE FIT
# P_verrucosa_Haplotype_3b
d.verrucosa.may <- Photodata %>% 
  filter(mtORF=="P_verrucosa_Haplotype_3b")%>% 
filter(Season=="May")

#fit 
verrucosa.fit.may <- nls_multstart(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.verrucosa.may,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

verrucosa.fit.may

#generate the predicted data
verrucosa_new_data.may <- data.frame(temp = seq(min(d.verrucosa.may$temp), max(d.verrucosa.may$temp), 0.5))
verrucosa.preds.may <- augment(verrucosa.fit.may, newdata = verrucosa_new_data.may)

#calculate TPC parameters
verrucosa.TCP.res.may <- calc_params(verrucosa.fit.may) %>%
  mutate_all(round, 2)   # round 

verrucosa.TCP.res.may 

# refit model using nlsLM
verrucosa.fit_nlsLM.may <- minpack.lm::nlsLM(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.verrucosa.may,
                        start = coef(verrucosa.fit.may),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.verrucosa.may)))

# bootstrap using case resampling
verrucosa.boot1.may <- Boot(verrucosa.fit_nlsLM.may, method = 'case')

# look at the data
head(verrucosa.boot1.may$t)


# create predictions of each bootstrapped model
verrucosa.boot1_preds.may <- verrucosa.boot1.may$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.verrucosa.may$temp), max(d.verrucosa.may$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
verrucosa.boot1_conf_preds.may <- group_by(verrucosa.boot1_preds.may, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
verrucosa.CI.plot.May <- ggplot() +
  geom_line(aes(temp, .fitted), verrucosa.preds.may, col = 'cyan') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds.may, fill = 'cyan', alpha = 0.3) +
  geom_point(aes(temp, transformed.rate), d.verrucosa.may, size = 2, alpha = 0.5,col = 'cyan') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Log 10 +1 Rate (µmol O2/cm2/h)')
verrucosa.CI.plot.May

```
#VERRUCOSA December
```{r}
# verrucosa CURVE FIT
# P_verrucosa_Haplotype_3b
d.verrucosa.dec <- Photodata %>% 
  filter(mtORF=="P_verrucosa_Haplotype_3b")%>% 
filter(Season=="Dec")

#fit 
verrucosa.fit.dec <- nls_multstart(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                                                     data = d.verrucosa.dec,
                                                     iter = 500,
                                                     start_lower = start_vals - 1,
                                                     start_upper = start_vals + 1,
                                                     lower = low_lims,
                                                     upper = upper_lims,
                                                     supp_errors = 'Y')

verrucosa.fit.dec

#generate the predicted data
verrucosa_new_data.dec <- data.frame(temp = seq(min(d.verrucosa.dec$temp), max(d.verrucosa.dec$temp), 0.5))
verrucosa.preds.dec <- augment(verrucosa.fit.dec, newdata = verrucosa_new_data.dec)

#calculate TPC parameters
verrucosa.TCP.res.dec <- calc_params(verrucosa.fit.dec) %>%
  mutate_all(round, 2)   # round 

verrucosa.TCP.res.dec 

# refit model using nlsLM
verrucosa.fit_nlsLM.dec <- minpack.lm::nlsLM(transformed.rate~sharpeschoolhigh_1981(temp = temp, r_tref,e,eh,th, tref = 28),
                        data = d.verrucosa.dec,
                        start = coef(verrucosa.fit.dec),
                        lower = low_lims,
                        upper = upper_lims,
                        weights = rep(1, times = nrow(d.verrucosa.dec)))

# bootstrap using case resampling
verrucosa.boot1.dec <- Boot(verrucosa.fit_nlsLM.dec, method = 'case')

# look at the data
head(verrucosa.boot1.dec$t)


# create predictions of each bootstrapped model
verrucosa.boot1_preds.dec <- verrucosa.boot1.dec$t %>%
  as.data.frame() %>%
  drop_na() %>%
  mutate(iter = 1:n()) %>%
  group_by_all() %>%
  do(data.frame(temp = seq(min(d.verrucosa.dec$temp), max(d.verrucosa.dec$temp), length.out = 100))) %>%
  ungroup() %>%
  mutate(pred = sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref = 28))

# calculate bootstrapped confidence intervals
verrucosa.boot1_conf_preds.dec <- group_by(verrucosa.boot1_preds.dec, temp) %>%
  summarise(conf_lower = quantile(pred, 0.025),
            conf_upper = quantile(pred, 0.975)) %>%
  ungroup()

# plot bootstrapped CIs
verrucosa.CI.plot.dec <- ggplot() +
  geom_line(aes(temp, .fitted), verrucosa.preds.dec, col = 'cyan') +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds.dec, fill = 'cyan', alpha = 0.3) +
  geom_point(aes(temp, transformed.rate), d.verrucosa.dec, size = 2, alpha = 0.5,col = 'cyan') +
  theme_bw(base_size = 12) +
  labs(x = 'Temperature (ºC)',
       y = 'Log 10 +1 Rate (µmol O2/cm2/h)')
verrucosa.CI.plot.dec

```
#Plot all curves together
```{r}
 #set plot colors
cols <- c("grandis.May"="orange","grandis.Dec"="orangered",  "verrucosa.May"="aquamarine",  "verrucosa.Dec"="cyan")
  
# plot data and model fit
TPC.plot <- ggplot(data=Photodata, aes(x=temp)) +
   geom_point(aes(temp, transformed.rate, color="grandis.May"), d.grandis.may, size = 2, alpha = 0.5) +
   geom_point(aes(temp, transformed.rate, color="grandis.Dec"), d.grandis.dec, size = 2, alpha = 0.5) +
   geom_point(aes(temp, transformed.rate, color="verrucosa.May"), d.verrucosa.may, size = 2, alpha = 0.5) +
   geom_point(aes(temp, transformed.rate, color="verrucosa.Dec"), d.verrucosa.dec, size = 2, alpha = 0.5) +
   geom_line(aes(temp, .fitted), grandis.preds.may, col = 'orange', size=2,linetype="dotted") +
   geom_line(aes(temp, .fitted), grandis.preds.dec, col = 'orangered', size=2) +
   geom_line(aes(temp, .fitted), verrucosa.preds.may, col = "aquamarine", size=2,linetype="dotted") +
   geom_line(aes(temp, .fitted), verrucosa.preds.dec, col = "cyan", size=2) +
   geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds.may, fill = "orange", alpha = 0.3) +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), grandis.boot1_conf_preds.dec, fill = "orangered", alpha = 0.3) +

  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds.may, fill = 'aquamarine', alpha = 0.3) +
  geom_ribbon(aes(temp, ymin = conf_lower, ymax = conf_upper), verrucosa.boot1_conf_preds.dec, fill = 'cyan', alpha = 0.3) +
   xlim(21,37)+
  #ylim(0,0.45)+
   scale_x_continuous(breaks=c(22,24,26,28,30,32,34,36))+
   theme_bw(base_size = 12) +
   scale_colour_manual(name="Morphology",values=cols)+
   theme(
         panel.border = element_blank(), panel.grid.major = element_blank(),
         panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))+
   labs(x = 'Temperature (ºC)',
        y = expression("Rate"~µmol~O[2] ~cm^{-2}~h^{-1}))

TPC.plot 

ggsave("output/Dec_TPC_SharpSchool_Photosynthesis.pdf", TPC.plot, dpi=300, w=8, h=8, units="in")
ggsave("output/Dec_TPC_SharpSchool_Photosynthesis.jpg", TPC.plot, dpi=300, w=8, h=8, units="in")
```

# Bootstrap CI for all TPC parameters
```{r}
broom::tidy(grandis.fit_nlsLM.may)
broom::tidy(grandis.fit_nlsLM.dec)
broom::tidy(verrucosa.fit_nlsLM.may)
broom::tidy(verrucosa.fit_nlsLM.dec)
```
#GRANDIS May
```{r}
#calculate all the TPC parameters
grandis.extra_params.may <- calc_params(grandis.fit_nlsLM.may) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
grandis.ci_extra_params.may <- Boot(grandis.fit_nlsLM.may, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(grandis.fit_nlsLM.may)), R = 200, method = 'case') %>%
  confint(., method = 'perc') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')

#join the parameters and CIs  
grandis.ci_extra_params.may <- left_join(grandis.ci_extra_params.may, grandis.extra_params.may)
grandis.ci_extra_params.may$Treatment <- "P. grandis May"
```
#GRANDIS December
```{r}
#calculate all the TPC parameters
grandis.extra_params.dec <- calc_params(grandis.fit_nlsLM.dec) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
grandis.ci_extra_params.dec <- Boot(grandis.fit_nlsLM.dec, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(grandis.fit_nlsLM.dec)), R = 200, method = 'case') %>%
  confint(., method = 'perc') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')

#join the parameters and CIs  
grandis.ci_extra_params.dec <- left_join(grandis.ci_extra_params.dec, grandis.extra_params.dec)
grandis.ci_extra_params.dec$Treatment <- "P. grandis December"
```


#VERRUCOSA May
```{r}
#calculate all the TPC parameters
verrucosa.extra_params.may <- calc_params(verrucosa.fit_nlsLM.may) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
verrucosa.ci_extra_params.may <- Boot(verrucosa.fit_nlsLM.may, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(verrucosa.fit_nlsLM.may)), R = 200, method = 'case') %>%
  confint(., method = 'perc') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')
  
#join the parameters and CIs  
verrucosa.ci_extra_params.may <- left_join(verrucosa.ci_extra_params.may, verrucosa.extra_params.may)
verrucosa.ci_extra_params.may$Treatment <- "P. verrucosa May"
```

#VERRUCOSA December
```{r}
#calculate all the TPC parameters
verrucosa.extra_params.dec <- calc_params(verrucosa.fit_nlsLM.dec) %>%
  pivot_longer(everything(), names_to =  'param', values_to = 'estimate')

#calculate CIs for all the TPC parameters
verrucosa.ci_extra_params.dec <- Boot(verrucosa.fit_nlsLM.dec, f = function(x){unlist(calc_params(x))}, labels = names(calc_params(verrucosa.fit_nlsLM.dec)), R = 200, method = 'case') %>%
  confint(., method = 'perc') %>%
  as.data.frame() %>%
  rename(conf_lower = 1, conf_upper = 2) %>%
  rownames_to_column(., var = 'param') %>%
  mutate(method = 'case bootstrap')
  
#join the parameters and CIs  
verrucosa.ci_extra_params.dec <- left_join(verrucosa.ci_extra_params.dec, verrucosa.extra_params.dec)
verrucosa.ci_extra_params.dec$Treatment <- "P. verrucosa December"
```

#Join Morphology estimates and CIs
```{r}
All_params <- rbind(grandis.ci_extra_params.may, grandis.ci_extra_params.dec,  verrucosa.ci_extra_params.may,  verrucosa.ci_extra_params.dec)
All_params <- All_params %>% 
 mutate_if(is.numeric, round, 2)

#order levels
All_params$Treatment <- factor(All_params$Treatment, levels = c("P. grandis May", "P. verrucosa May", "P. grandis December", "P. verrucosa December"))
All_params$Treatment


#Plot all of the estimates
estimate.plots <- ggplot(All_params, aes(Treatment, estimate, color=Treatment)) +
  geom_point(size = 2) +
  scale_color_manual(name="Treatment", values=c("orange","aquamarine", "orangered","cyan"))+
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  facet_wrap(~param, scales = 'free_y') +
  scale_x_discrete('')

estimate.plots

#filter to only the most relavent and well characterized parameters
All_params <- All_params %>% 
  filter(!param=="ctmin") %>%
  filter(!param=="eh") %>%
  #filter(!param=="rmax") %>%
  filter(!param=="skewness") %>%
  filter(!param=="thermal_safety_margin") %>%
  #filter(!param=="topt") %>%
  filter(!param=="thermal_tolerance") %>%
  filter(!param=="q10") %>%
  filter(!param=="e") %>%
  filter(!param=="breadth")
  
#view estimate plots
estimate.plots <- ggplot(All_params, aes(Treatment, estimate, color=Treatment)) +
  geom_point(size = 2) +
  scale_color_manual(name="Treatment", values=c("orange","aquamarine", "orangered","cyan"))+
  geom_linerange(aes(ymin = conf_lower, ymax = conf_upper)) +
  theme_bw() +
  labs(y = NULL, color=NULL)+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        legend.position = "right",
        strip.background = element_blank(), 
        strip.placement = "outside") +
  facet_wrap(~param, scales = 'free_y', nrow=1)+
             #labeller = as_labeller(c(e = "e (Energy)", eh = " eh (Energy)", rmax= "Rmax (~nmol~O[2] ~larva^{-1}~min^{-1})",topt="Topt (Temperature °C)")), strip.position = "left") +
  theme(legend.position = "none")+
  scale_x_discrete('')

estimate.plots

ggsave("output/Dec_TPC_estimates_SharpSchool_Photosynthesis.pdf", estimate.plots, dpi=300, w=6, h=2, units="in")
ggsave("output/Dec_TPC_estimates_SharpSchool_Photosynthesis.jpg", estimate.plots, dpi=300, w=6, h=2, units="in")
``` 

#Plot Curve and Estimate Output
```{r}
#generate a combined figure of TPCs and estimate plots
figure <- ggarrange(TPC.plot , estimate.plots,
                    labels = c("A", "B"),
                    ncol = 1, nrow = 2,
                    heights=c(1,0.5))
figure

ggsave("output/MayDec_Photosynthesis_TPC_and_estimates.pdf", figure, dpi=300, w=6, h=8, units="in")
ggsave("output/MayDec_Photosynthesis_TPC_and_estimates.jpg", figure, dpi=300, w=6, h=8, units="in")

```
