==> Ranalysis/scripts/5_phys_surface_area.Rmd <==
---
title: "Surface Area Calculations from Wax Dipping"
author: "HM Putnam"
date: "20220805"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## install packages if you dont already have them
if (!require("tidyverse")) install.packages("tidyverse")

# load packages
library(tidyverse)
```

#use wooden dowel standards to calculate surface area curve 
```{r}

#load wax data
wax.data <- read.csv("data/May_TPC_Phys_surface_area_data.csv", header=TRUE)

wax.data$delta.mass.g <- wax.data$weight2.g-wax.data$weight1.g
stnds <- subset(wax.data, Sample=="Standard")
stnds <- stnds[-1,] # the largest data point was an artifact outlier. It was removed, as it had bubbles of air escape the wooden sphere, leaving gaps in the wax

#calculate the surface area of the spherical standards from the diameter
stnds$rad <- stnds$Diameter/2
stnds$surface.area.cm2 <- 4*pi*(stnds$rad)^2

# calculate the curve coefficients for slope and intercept to apply as the standard
stnd.curve <- lm(surface.area.cm2~delta.mass.g, data=stnds)
plot(surface.area.cm2~delta.mass.g, data=stnds)
stnd.curve$coefficients
summary(stnd.curve)$r.squared 


```

#May 2022 Data
```{r}
#Calculate surface area using the standard curve
smpls <- subset(wax.data, Sample=="Coral")
smpls$surface.area.cm2 <- stnd.curve$coefficients[2] * smpls$delta.mass.g + stnd.curve$coefficients[1]

#select the samples only
smpls <- smpls %>%
  select(-Sample, -Diameter)

smpls %>%
  count(fragment_ID) %>% arrange(n)

#check the range to make sure your samples fall within the range of the standards
range(smpls$surface.area.cm2)
hist(smpls$surface.area.cm2)
range(stnds$surface.area.cm2)

sa_may <- smpls
sa_may$Timepoint <- "May"
write_csv(sa_may, "data/May_TPC_Phys_surface.area.calc.csv")

```

#December 2022 Data
```{r}
#Calculate surface area using the standard curve
wax.data.2 <- read.csv("data/Dec_TPC_Phys_surface_area_data.csv", header=TRUE)
smpls.2 <- subset(wax.data.2, Sample=="Coral")
smpls.2$delta.mass.g <- smpls.2$weight2.g-smpls.2$weight1.g
smpls.2$surface.area.cm2 <- stnd.curve$coefficients[2] * smpls.2$delta.mass.g + stnd.curve$coefficients[1]

#select the samples only
smpls.2 <- smpls.2 %>%
  select(-Sample, -Diameter)

smpls.2 %>%
  count(fragment_ID) %>% arrange(n)

#check the range to make sure your samples fall within the range of the standards
range(smpls.2$surface.area.cm2, na.rm = T)
hist(smpls.2$surface.area.cm2)
range(stnds$surface.area.cm2)

sa_dec <- smpls.2
sa_dec$Timepoint <- "Dec"
write_csv(sa_dec, "data/Dec_TPC_Phys_surface.area.calc.csv")

```


#Save the output for use in normilzation for phys assays
```{r}

#surface.area.cm2 <- rbind(sa_may, sa_dec)

#surface.area.cm2$ID <- paste0(surface.area.cm2$fragment_ID, surface.area.cm2$Timepoint)

#surface.area.cm2 %>%
#  select(fragment_ID, Timepoint, ID, surface.area.cm2)%>%
#    write_csv("output/TPC_Phys.surface.area.calc.csv")
```
