==> Ranalysis/scripts/7_chlorophyll.Rmd <==
---
title: "Chlorophyll analysis"
author: "HM Putnam, AS Huffmyer"
date: '20220324'
output:
  pdf_document: default
  html_document: default
edited by: DM Becker-Polinski
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

# load packages
library(plotrix)
library(tidyverse)
library(Hmisc)
library(performance)
library(ggpubr)
library(rstatix)
library(see)

```

# May
### Import data
```{r}
# Define function to read in chl data
read_chl <- function(file) {
  chl_data <- read_csv(file, skip = 24, n_max = 24) %>%
    select(-1) %>%
    magrittr::set_colnames(c("row", 1:12, "wavelength")) %>%
    fill(row) %>%
    gather("col", "absorbance", -wavelength, -row) %>%
    unite("well", c(row, col), sep = "")
}

# List chlorophyll data files
chl_path <- "/Users/hputnam/MyProjects/Poc_ID/Ranalysis/data/chl/May/"                                        # Path to chlorophyll data directory
all_chl_files <- list.files(path = chl_path, pattern = "_Chl_data.csv")          # List all files in directory
chl_platemaps <- list.files(path = chl_path, pattern = "Chl_data_platemap")       # List platemap files
chl_data_files <- setdiff(all_chl_files, chl_platemaps)                  # List absorbance data files


# Read in all files into tibble
df <- tibble(file = chl_data_files) %>%
  mutate(platemap = map(file, ~ read_csv(paste0(chl_path, tools::file_path_sans_ext(.), "_platemap.csv"))),
         chl_data = map(file, ~ read_chl(paste0(chl_path, .))))

# Merge platemap and data for each plate
df <- df %>%
  mutate(merged = map2(platemap, chl_data, ~ right_join(.x, .y)))


# Calculate chlorophyll concentrations

# average all technical replicates for each plate/sample/wavelength, including all acetone blanks together (per plate)
df <- df %>%
  unnest(merged) %>%
  filter(!is.na(fragment_ID)) %>%                         # remove empty wells (fragment_ID is NA)
  group_by(file, fragment_ID, wavelength) %>%
  summarise(n = n(), mean_abs = mean(absorbance)) %>%
  spread(wavelength, mean_abs)

# get the acetone blank 750 absorbace for each file (i.e., plate), and subtract from 630 and 663 values for each sample
df <- df %>%
  group_by(file) %>%
  mutate(blank750 = `750`[fragment_ID == "BK"]) %>%
  ungroup() %>%
  mutate(adj630 = `630` - blank750,
         adj663 = `663` - blank750)

# calculate chla and chlc2 values based on equations from Jeffrey and Humphrey 1975
# units µg/ml
#path length adjustment = 0.6 

df.may <- df %>%
  mutate(chla.ug.ml = (11.43 * adj663)/0.6 - (0.64 * adj630)/0.6,
        chlc2.ug.ml = (27.09 * adj630)/0.6 - (3.63 * adj663)/0.6)

df.may$Timepoint <- "May"

```

# Dec
### Import data
```{r}

# List chlorophyll data files
chl_path <- "/Users/hputnam/MyProjects/Poc_ID/Ranalysis/data/chl/Dec/"                                        # Path to chlorophyll data directory
all_chl_files <- list.files(path = chl_path, pattern = "_Chl_data.csv")          # List all files in directory
chl_platemaps <- list.files(path = chl_path, pattern = "Chl_data_platemap")       # List platemap files
chl_data_files <- setdiff(all_chl_files, chl_platemaps)                  # List absorbance data files


# Read in all files into tibble
df <- tibble(file = chl_data_files) %>%
  mutate(platemap = map(file, ~ read_csv(paste0(chl_path, tools::file_path_sans_ext(.), "_platemap.csv"))),
         chl_data = map(file, ~ read_chl(paste0(chl_path, .))))

# Merge platemap and data for each plate
df <- df %>%
  mutate(merged = map2(platemap, chl_data, ~ right_join(.x, .y)))


# Calculate chlorophyll concentrations

# average all technical replicates for each plate/sample/wavelength, including all acetone blanks together (per plate)
df <- df %>%
  unnest(merged) %>%
  filter(!is.na(fragment_ID)) %>%                         # remove empty wells (fragment_ID is NA)
  group_by(file, fragment_ID, wavelength) %>%
  summarise(n = n(), mean_abs = mean(absorbance)) %>%
  spread(wavelength, mean_abs)

# get the acetone blank 750 absorbace for each file (i.e., plate), and subtract from 630 and 663 values for each sample
df <- df %>%
  group_by(file) %>%
  mutate(blank750 = `750`[fragment_ID == "BK"]) %>%
  ungroup() %>%
  mutate(adj630 = `630` - blank750,
         adj663 = `663` - blank750)

# calculate chla and chlc2 values based on equations from Jeffrey and Humphrey 1975
# units µg/ml
#path length adjustment = 0.6 

df.dec <- df %>%
  mutate(chla.ug.ml = (11.43 * adj663)/0.6 - (0.64 * adj630)/0.6,
        chlc2.ug.ml = (27.09 * adj630)/0.6 - (3.63 * adj663)/0.6)

df.dec$Timepoint <- "Dec"

chl <- rbind(df.may, df.dec) 

chl$ID <- paste0(chl$fragment_ID, chl$Timepoint)
```


# Normalize to surface area
```{r}
# Load homogenate volume
homog.vol <- read_csv("data/homogenate_vols.csv") %>%
  select(fragment_ID, homog_vol_ml, Timepoint)

homog.vol$ID <- paste0(homog.vol$fragment_ID, homog.vol$Timepoint)

chl.data <- left_join(chl, homog.vol, by="ID")

# remove blanks and NAs
chl.data <- filter(chl.data, !ID %in% c("NA", "BK"))

# Load Surface area data
sa <- read.csv("output/TPC_Phys.surface.area.calc.csv")

# Coral sample metadata
metadata <- read_csv("data/metadata_POC_TPC.csv")

# Join homogenate volumes and surface area with sample metadata
chl.data <- dplyr::inner_join(
   chl.data,
   dplyr::select(metadata, -any_of(names(chl.data)), ID),
   by = "ID"
)

chl.data <- dplyr::inner_join(
   chl.data,
   dplyr::select(sa, -any_of(names(chl.data)), ID),
   by = "ID"
)


# Multiply chlorophyll by the homogenate volume and divide by surface area
chl.data <- chl.data %>%
  mutate(chla.ug.cm2 = chla.ug.ml * homog_vol_ml / surface.area.cm2,
         chlc2.ug.cm2 = chlc2.ug.ml * homog_vol_ml / surface.area.cm2)

# remove blanks and NAs
chl.data <- filter(chl.data, !ID %in% c("NA", "BK"))

# write chlorophyll data to file
chl.data %>%
  select(ID, Timepoint.x, Morphology,chla.ug.cm2, chlc2.ug.cm2) %>%
  filter(!is.na(chla.ug.cm2))%>%
  filter(!is.na(chlc2.ug.cm2))
 # write_csv(path = "output/TPC_Phys_chlorophyll.csv")
```

#filter to Hap1a and Hap3b only and the same colonies for both times
```{r}
chl.sub<-chl.data %>% 
  filter(mtORF =="Haplotype_1a" | mtORF =="P verrucosa Haplotype 3b")

chl.sub.unique <- chl.sub %>% 
  group_by(fragment_ID.x) %>% 
  filter(n_distinct(Timepoint.x) == 2)

```

#Statistical testing chla
```{r}


# 2 way ANOVA
## Untransformed
model  <- lm(chla.ug.cm2 ~ mtORF*Timepoint.x, data = chl.sub.unique)
check_model(model)
check_outliers(model)
check_normality(model)

## log10 Transformed
model.1  <- lm(log10(chla.ug.cm2) ~ mtORF*Timepoint.x, data = chl.sub.unique)
check_model(model.1)
check_outliers(model.1)
check_normality(model.1)

model.chla  <- lm(log10(chla.ug.cm2) ~ mtORF*Timepoint.x, data = chl.sub.unique)
anova(model.chla )
stat.labels1 <- round(anova(model.chla), 5)

```

#Statistical testing chlc2
```{r}

## Outlier Detection

# 2 way ANOVA
## Untransformed
model  <- lm(chlc2.ug.cm2 ~ mtORF*Timepoint.x, data = chl.sub.unique)
check_model(model)
check_outliers(model)
check_normality(model)

## log10 Transformed
model.1  <- lm(log10(chlc2.ug.cm2) ~ mtORF*Timepoint.x, data = chl.sub.unique)
check_model(model.1)
check_outliers(model.1)
check_normality(model.1)

model.chlc2  <- lm(log10(chlc2.ug.cm2) ~ mtORF*Timepoint.x, data = chl.sub.unique)
anova(model.chlc2)
stat.labels2 <- round(anova(model.chlc2), 5)

```



```{r}
#summarize mean and sem by site, species, and partner and plot

level_order <- c('May', 'Dec') 
chl.sub.unique$group <- paste0(chl.sub.unique$Timepoint, " ", chl.sub.unique$mtORF)

# Plot all data points with mean ± se
chla_plot <- chl.sub.unique %>%
  ggplot(aes(x = factor(Timepoint.x, level = level_order), y = chla.ug.cm2, color = group, shape=mtORF, group = mtORF)) +
  labs(x = "", y = expression("Chl-a µg"~cm^{-2})) +
  geom_jitter(width = 0.3) +                                            # Plot all points
  stat_summary(fun.data = mean_cl_normal, fun.args = list(mult = 1),    # Plot standard error
               geom = "errorbar", width = 0.2) +
  stat_summary(fun = mean, geom = "point", size=3) +           # Plot mean
    stat_summary(aes(x = Timepoint, group = mtORF), fun = mean, geom = "line", alpha=0.2) +
  scale_color_manual(name="group", values=c( "orangered","cyan","orange", "aquamarine"))+
  geom_text(x=2, y=6.3, label=paste("Species=",stat.labels1$`Pr(>F)`[1]), size=3, color=ifelse(stat.labels1$`Pr(>F)`[1] < 0.05, "black", "darkgray")) + 
  geom_text(x=2, y=6.1, label=paste("Time=",stat.labels1$`Pr(>F)`[2]), size=3, color=ifelse(stat.labels1$`Pr(>F)`[2] < 0.05, "black", "darkgray")) + 
  geom_text(x=2, y=5.9, label=paste("Species x Time=",stat.labels1$`Pr(>F)`[3]), size=3, color=ifelse(stat.labels1$`Pr(>F)`[3] < 0.05, "black", "darkgray")) +
  theme(axis.text.y = element_text(size=14, face="bold", colour = "black"),axis.text.x = element_text(size=14, face="bold", colour = "black")) +
  theme_classic(base_size = 14)+
  theme(legend.position='none')
chla_plot 


```

```{r}

chl.sub.unique$group <- paste0(chl.sub.unique$Timepoint, " ", chl.sub.unique$mtORF)


# Plot all data points with mean ± se
chlc2_plot <- chl.sub.unique %>%
  ggplot(aes(x = factor(Timepoint.x, level = level_order), y = chlc2.ug.cm2, color = group, shape=mtORF, group = mtORF)) +
  labs(x = "", y = expression("Chl-c2 µg"~cm^{-2})) +
  geom_jitter(width = 0.3) +                                            # Plot all points
  stat_summary(fun.data = mean_cl_normal, fun.args = list(mult = 1),    # Plot standard error
               geom = "errorbar", width = 0.2) +
  stat_summary(fun = mean, geom = "point", size=3) +           # Plot mean
  stat_summary(aes(x = Timepoint, group = mtORF), fun = mean, geom = "line", alpha=0.2) +
  scale_color_manual(name="group", values=c( "orangered","cyan","orange", "aquamarine"))+
  geom_text(x=2, y=1.7, label=paste("Species=",stat.labels2$`Pr(>F)`[1]), size=3, color=ifelse(stat.labels2$`Pr(>F)`[1] < 0.05, "black", "darkgray")) + 
  geom_text(x=2, y=1.63, label=paste("Time=",stat.labels2$`Pr(>F)`[2]), size=3, color=ifelse(stat.labels2$`Pr(>F)`[2] < 0.05, "black", "darkgray")) + 
  geom_text(x=2, y=1.56, label=paste("Species x Time=",stat.labels2$`Pr(>F)`[3]), size=3, color=ifelse(stat.labels2$`Pr(>F)`[3] < 0.05, "black", "darkgray")) +
  theme(axis.text.y = element_text(size=14, face="bold", colour = "black"),axis.text.x = element_text(size=14, face="bold", colour = "black")) +
  theme_classic(base_size = 14)+
  theme(legend.position='none')
chlc2_plot 

```


```{r}
 (chla_plot+chlc2_plot) 
  ggsave("output/Poc_TPC_Chl_1a3b.pdf", plot = last_plot(), width = 10, height = 4)
  ggsave("output/Poc_TPC_Chl_1a3b.jpg", plot = last_plot(), width = 10, height = 4)
```
